requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==11.0.0
psycopg2-binary==2.9.9
